Remove testing configurations. These don't actually add new functionality - We can use the Development and Staging configurations as long as we set the values correctly and correctly isolate the environment - Move the empty_environment and fake_prod_env fixtures to conftest.py to give all tests access if they need it
